Add files via upload
This is version 1 of the production calculator. You can only calculate one tank at a time, and it is a stand alone file anyone can download onto their computer and pin to their taskbar or create a desktop shortcut. You do not need to download anything on your computer to use this.